requirements.txt
absl-py==0.12.0
addict==2.4.0
appdirs==1.4.4
apptools==5.1.0
argon2-cffi==20.1.0
astroid==2.12.9
async-generator==1.10
attrs==20.3.0
black==20.8b1
bleach==3.3.0
boto3==1.17.78
botocore==1.20.78
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chumpy==0.70
click==7.1.2
colorama==0.4.4
configobj==5.0.6
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
Deprecated==1.2.12
dill==0.3.5.1
easydict==1.9
entrypoints==0.3
envisage==5.0.0
filelock==3.0.12
flatbuffers==1.12
freetype-py==2.2.0
future==0.18.2
fvcore==0.1.5.post20220305
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.10.1
importlib-resources==5.1.2
iopath==0.1.9
ipywidgets==7.6.3
ipykernel==5.3.4
ipython==7.22.0
ipython-genutils==0.2.0
isort==5.10.1
jax==0.2.12
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.7.1
libaio==0.8.3
lmdb==1.2.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.7.0
mistune==0.8.4
mkl-fft==1.3.0
mujoco==2.3.3
MulticoreTSNE==0.1
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.19.2
oauthlib==3.1.0
olefile==0.46
open3d==0.12.0
opencv-contrib-python-headless==4.5.1.48
opencv-python-headless==4.5.1.48
opendr==0.78
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0 
pickleshare==0.7.5
Pillow==8.0.1
platformdirs==2.5.2
plyfile==0.7.4
portalocker==2.4.0
prettytable==2.1.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pyembree==0.1.6
pyface==7.3.0
pyglet==1.5.16
Pygments==2.8.1
pylint==2.15.2
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrender==0.1.43
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
qrcode==6.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.4.2
sacremoses==0.0.45
scikit-learn==0.24.1
scikit-video==1.1.11
scipy==1.5.2
Send2Trash==1.5.0
six== 1.15.0 
sklearn==0.0
smmap==4.0.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
torch-tb-profiler==0.2.1
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
traits==6.2.0
traitsui==7.1.1
transformers==4.5.1
transforms3d==0.3.1
trimesh==3.9.30
typed-ast==1.4.3
typing-extensions==4.3.0
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.14.1
yacs==0.1.8
zipp==3.4.1

liegroups @ git+https://github.com/utiasSTARS/liegroups.git
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@v0.7.2
manotorch @ git+https://github.com/lixiny/manotorch.git@63037ea
chamfer-distance @ git+https://github.com/KailinLi/chamfer_distance.git
manopth @ git+https://github.com/hassony2/manopth.git
